move dynamic filter section component into lib
