+++ client/web-sveltekit/src/lib/search/dynamicFilters/Section.svelte
